# ApiBackendOrchestrator/architecture/api-specification.yaml
openapi: 3.0.3
info:
  title: ContractPro API — Backend Orchestrator
  description: |
    REST API оркестратора ContractPro — единая точка входа для frontend-приложений и внешних интеграций.
    Оркестратор координирует работу доменных сервисов (DM, DP, OPM, UOM) и предоставляет
    агрегированные данные для пользователей.

    ## Аутентификация
    Все endpoints (кроме /auth/*) требуют JWT Bearer token в заголовке Authorization.
    JWT содержит claims: user_id, organization_id, role.

    ## Формат ошибок
    Все ошибки возвращаются в едином формате ErrorResponse с кодом ошибки, сообщением на русском языке
    и рекомендацией следующего шага для пользователя (NFR-5.2).

    ## Real-time обновления
    Статусы обработки доставляются через SSE endpoint GET /api/v1/events/stream.
  version: 1.0.0
  contact:
    name: ContractPro Backend Team
servers:
  - url: /api/v1
    description: Orchestrator API v1
security:
  - BearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Аутентификация пользователя
      description: Проксирование в UOM. Возвращает JWT access token и refresh token.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Неверные учётные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Обновление access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Новый access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Невалидный или отозванный refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Выход из системы
      description: Отзывает refresh token в UOM.
      operationId: logout
      responses:
        '204':
          description: Выход выполнен
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== Contracts ====================
  /contracts/upload:
    post:
      tags: [Contracts]
      summary: Загрузка нового договора
      description: |
        Загрузка PDF-файла договора. Оркестратор:
        1. Валидирует файл (размер, формат).
        2. Загружает в Object Storage.
        3. Создаёт документ и версию в DM.
        4. Публикует команду на обработку в DP.
        Возвращает 202 Accepted — обработка асинхронная.
      operationId: uploadContract
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, title]
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF-файл договора (макс. 20 МБ)
                title:
                  type: string
                  minLength: 1
                  maxLength: 500
                  description: Название договора
      responses:
        '202':
          description: Договор принят в обработку
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: Файл слишком большой
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Неподдерживаемый формат файла
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /contracts:
    get:
      tags: [Contracts]
      summary: Список договоров организации
      description: |
        Возвращает постраничный список договоров организации. Каждый
        `ContractSummary` несёт агрегаты текущей версии — `contract_type`,
        `risk_level`, `risk_counts` (ORCH-TASK-056), — собираемые Orchestrator из
        артефактов Document Management ОДНИМ батч-запросом на страницу (без N+1,
        см. high-architecture.md §8.7 «Список документов с агрегацией»).

        Серверная фильтрация и сортировка по типу/риску/статусу/периоду
        (`risk_level`, `contract_type`, `processing_status`, `date_from`,
        `date_to`, `sort`, `order`) делегируются DM (источник истины), чтобы
        `total` и пагинация оставались корректными. Все новые параметры
        опциональны; обратная совместимость сохранена.

        Доступность агрегации зависит от read-контракта DM
        `GET /documents?include=analysis` (ASSUMPTION-ORCH-17) и управляется
        флагом `ORCH_CONTRACTS_LIST_ANALYSIS_ENABLED`. Пока флаг выключен,
        агрегатные поля приходят как `null`, а передача новых фильтров/сортировки
        отклоняется с 400 (fail-safe: список не возвращает «как бы
        отфильтрованные», но фактически неотфильтрованные данные).
      operationId: listContracts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/DocumentStatus'
          description: Фильтр по статусу документа (DocumentStatus — ACTIVE/ARCHIVED/DELETED)
        - name: search
          in: query
          schema:
            type: string
            maxLength: 200
          description: Поиск по названию (substring match)
        - name: risk_level
          in: query
          schema:
            type: string
            enum: [high, medium, low]
          description: |
            Фильтр по агрегированному уровню риска текущей версии (одиночное
            значение). Учитываются только договоры с READY-версией.
        - name: contract_type
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ContractType'
          description: |
            Фильтр по типу договора (английский LIC enum). Можно указывать
            несколько значений (`?contract_type=LEASE&contract_type=SALE`) — они
            объединяются по ИЛИ.
        - name: processing_status
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/UserProcessingStatus'
          description: |
            Фильтр по статусу обработки текущей версии (UserProcessingStatus).
            Несколько значений объединяются по ИЛИ. Orchestrator разворачивает
            каждый user-статус в множество DM `artifact_status` (например
            `ANALYZING` → {ARTIFACTS_READY, ANALYSIS_IN_PROGRESS}). Значение
            `AWAITING_USER_INPUT` управляется Orchestrator и не поддерживается
            для DM-фильтрации в этой версии → 400 VALIDATION_ERROR.
        - name: date_from
          in: query
          schema:
            type: string
          description: |
            Нижняя граница периода по дате создания договора (`created_at`),
            включительно. ISO-8601 — `YYYY-MM-DD` или RFC3339. Чипы «Период»
            7д/30д/90д на Frontend задают `date_from`.
        - name: date_to
          in: query
          schema:
            type: string
          description: |
            Верхняя граница периода по `created_at`, включительно. ISO-8601.
            Должна быть не раньше `date_from` (иначе 400).
        - name: sort
          in: query
          schema:
            type: string
            enum: [date, title, risk]
          description: |
            Поле сортировки: `date` (по `created_at`), `title`, `risk`. Для
            `risk` порядок по уровню (high > medium > low); договоры без риска
            (null) всегда в конце независимо от направления.
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
          description: Направление сортировки (по умолчанию desc на стороне DM).
      responses:
        '200':
          description: Список договоров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractList'
              example:
                items:
                  - contract_id: 550e8400-e29b-41d4-a716-446655440000
                    title: Договор аренды №17
                    status: ACTIVE
                    current_version_number: 2
                    processing_status: READY
                    contract_type: LEASE
                    risk_level: high
                    risk_counts: {high: 3, medium: 2, low: 1}
                    created_at: '2026-01-15T10:00:00Z'
                    updated_at: '2026-01-16T09:30:00Z'
                  - contract_id: 660e8400-e29b-41d4-a716-446655440111
                    title: Договор поставки №42
                    status: ACTIVE
                    current_version_number: 1
                    processing_status: ANALYZING
                    contract_type: SUPPLY
                    risk_level: null
                    risk_counts: null
                    created_at: '2026-02-01T08:00:00Z'
                    updated_at: '2026-02-01T08:05:00Z'
                total: 2
                page: 1
                size: 20
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /contracts/stats:
    get:
      tags: [Contracts]
      summary: Агрегированная статистика договоров организации
      description: |
        Возвращает сводные счётчики договоров текущей организации, агрегированные
        по статусу обработки (UserProcessingStatus) текущей версии и по уровню
        риска готовых договоров. Источник истины — Document Management; Orchestrator
        только агрегирует (§8.3 high-architecture). Используется дашбордом
        (карточка «Сводка»: «проверено» = total, «в работе» = сумма незавершённых
        статусов). Скоупится по организации пользователя из JWT.
      operationId: getContractStats
      parameters:
        - name: include_archived
          in: query
          schema:
            type: boolean
            default: false
          description: |
            Включать ли архивированные договоры (status=ARCHIVED) в статистику.
            По умолчанию учитываются только активные (status=ACTIVE).
      responses:
        '200':
          description: |
            Статистика договоров. Ответ помечен `Cache-Control: private, max-age=30`
            (агрегат скоупится по организации и устойчив к небольшой устареваемости).
          headers:
            Cache-Control:
              schema:
                type: string
              description: 'private, max-age=30 — приватное краткосрочное кэширование.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractStats'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /contracts/{contract_id}:
    get:
      tags: [Contracts]
      summary: Детали договора
      description: Возвращает метаданные договора с текущей версией и user-friendly статусом.
      operationId: getContract
      parameters:
        - $ref: '#/components/parameters/ContractIdParam'
      responses:
        '200':
          description: Детали договора
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Contracts]
      summary: Удаление договора (soft delete)
      operationId: deleteContract
      parameters:
        - $ref: '#/components/parameters/ContractIdParam'
      responses:
        '200':
          description: Договор удалён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /contracts/{contract_id}/archive:
    post:
      tags: [Contracts]
      summary: Архивация договора
      operationId: archiveContract
      parameters:
        - $ref: '#/components/parameters/ContractIdParam'
      responses:
        '200':
          description: Договор архивирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  # ==================== Versions ====================
  /contracts/{contract_id}/versions:
    get:
      tags: [Versions]
      summary: Список версий договора
      operationId: listVersions
      parameters:
        - $ref: '#/components/parameters/ContractIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Список версий
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /contracts/{contract_id}/versions/upload:
    post:
      tags: [Versions]
      summary: Загрузка новой версии существующего договора
      description: |
        Загрузка новой редакции (RE_UPLOAD). Создаёт новую версию и запускает обработку.
      operationId: uploadNewVersion
      parameters:
        - $ref: '#/components/parameters/ContractIdParam'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF-файл новой версии
      responses:
        '202':
          description: Новая версия принята в обработку
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          description: Файл слишком большой
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /contracts/{contract_id}/versions/{version_id}:
    get:
      tags: [Versions]
      summary: Детали версии
      operationId: getVersion
      parameters:
        - $ref: '#/components/parameters/ContractIdParam'
        - $ref: '#/components/parameters/VersionIdParam'
      responses:
        '200':
          description: Детали версии с user-friendly статусом
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /contracts/{contract_id}/versions/{version_id}/status:
    get:
      tags: [Versions]
      summary: Статус обработки версии (polling fallback)
      description: Альтернатива SSE для получения текущего статуса обработки.
      operationId: getVersionStatus
      parameters:
        - $ref: '#/components/parameters/ContractIdParam'
        - $ref: '#/components/parameters/VersionIdParam'
      responses:
        '200':
          description: Текущий статус обработки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /contracts/{contract_id}/versions/{version_id}/recheck:
    post:
      tags: [Versions]
      summary: Повторная проверка версии (UR-9)
      description: |
        Создаёт новую версию с origin_type=RE_CHECK (тот же файл, новый анализ).
      operationId: recheckVersion
      parameters:
        - $ref: '#/components/parameters/ContractIdParam'
        - $ref: '#/components/parameters/VersionIdParam'
      responses:
        '202':
          description: Повторная проверка запущена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Текущая версия ещё обрабатывается
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /contracts/{contract_id}/versions/{version_id}/confirm-type:
    post:
      tags: [Versions]
      summary: Подтверждение типа договора при низкой уверенности (UR-3 / FR-2.1.3)
      description: |
        Возобновляет обработку версии, остановленной в статусе AWAITING_USER_INPUT.
        Вызывается пользователем после получения SSE-события `type_confirmation_required`.

        Контракт:
        - Тело запроса: `{contract_type, confirmed_by_user: true}`. `contract_type` —
          либо `suggested_type` из события, либо одно из значений `alternatives[].contract_type`,
          либо произвольный валидный тип (контролируется LIC).
        - Версия должна находиться в статусе `AWAITING_USER_INPUT`. Иначе — 409
          с error_code `VERSION_NOT_AWAITING_INPUT`.
        - После успешного вызова Orchestrator публикует команду `UserConfirmedType`
          в LIC и переводит версию в `ANALYZING`.

        По истечении `ORCH_USER_CONFIRMATION_TIMEOUT` (default 24h) без подтверждения
        версия автоматически переводится в `FAILED` с error_code `USER_CONFIRMATION_TIMEOUT`
        (watchdog оркестратора).
      operationId: confirmContractType
      parameters:
        - $ref: '#/components/parameters/ContractIdParam'
        - $ref: '#/components/parameters/VersionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmTypeRequest'
      responses:
        '202':
          description: Подтверждение принято, обработка возобновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: |
            Невалидный contract_type — `error_code: INVALID_CONTRACT_TYPE` (не соответствует ни одному русскому лейблу UI, ни одному английскому LIC enum из 12-значного whitelist; см. event-catalog.md §1.3 для списка)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Версия не в статусе AWAITING_USER_INPUT (например, уже подтверждена, истёк таймаут или ещё анализируется)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== Results ====================
  /contracts/{contract_id}/versions/{version_id}/results:
    get:
      tags: [Results]
      summary: Агрегированные результаты проверки
      description: |
        Возвращает все результаты анализа для версии: риски, рекомендации, резюме,
        ключевые параметры, тип договора, сводная оценка. Фильтруется по роли пользователя.
        R-2 (бизнес-пользователь) видит только summary и aggregate_score.
      operationId: getResults
      parameters:
        - $ref: '#/components/parameters/ContractIdParam'
        - $ref: '#/components/parameters/VersionIdParam'
      responses:
        '200':
          description: Результаты проверки (или статус обработки, если не готовы)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResults'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /contracts/{contract_id}/versions/{version_id}/risks:
    get:
      tags: [Results]
      summary: Список рисков (UR-4, UR-5)
      operationId: getRisks
      parameters:
        - $ref: '#/components/parameters/ContractIdParam'
        - $ref: '#/components/parameters/VersionIdParam'
      responses:
        '200':
          description: Список рисков с пояснениями
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /contracts/{contract_id}/versions/{version_id}/summary:
    get:
      tags: [Results]
      summary: Краткое резюме (UR-7)
      operationId: getSummary
      parameters:
        - $ref: '#/components/parameters/ContractIdParam'
        - $ref: '#/components/parameters/VersionIdParam'
      responses:
        '200':
          description: Краткое резюме для бизнес-пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractSummaryResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /contracts/{contract_id}/versions/{version_id}/recommendations:
    get:
      tags: [Results]
      summary: Рекомендации формулировок (UR-6)
      operationId: getRecommendations
      parameters:
        - $ref: '#/components/parameters/ContractIdParam'
        - $ref: '#/components/parameters/VersionIdParam'
      responses:
        '200':
          description: Список рекомендаций
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== Comparison ====================
  /contracts/{contract_id}/compare:
    post:
      tags: [Comparison]
      summary: Запуск сравнения версий (FR-5.3.1)
      operationId: compareVersions
      parameters:
        - $ref: '#/components/parameters/ContractIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompareRequest'
      responses:
        '202':
          description: Сравнение запущено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompareResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /contracts/{contract_id}/versions/{base_version_id}/diff/{target_version_id}:
    get:
      tags: [Comparison]
      summary: Результат сравнения версий
      operationId: getDiff
      parameters:
        - $ref: '#/components/parameters/ContractIdParam'
        - name: base_version_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: target_version_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Результат сравнения
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionDiff'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Diff не найден (ещё не готов или не запускался)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== Export ====================
  /contracts/{contract_id}/versions/{version_id}/export/{format}:
    get:
      tags: [Export]
      summary: Экспорт отчёта (UR-10)
      operationId: exportReport
      parameters:
        - $ref: '#/components/parameters/ContractIdParam'
        - $ref: '#/components/parameters/VersionIdParam'
        - name: format
          in: path
          required: true
          schema:
            type: string
            enum: [pdf, docx]
      responses:
        '302':
          description: Redirect на presigned URL для скачивания
          headers:
            Location:
              schema:
                type: string
                format: uri
              description: Presigned S3 URL (TTL 5 минут)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Отчёт ещё не готов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== Feedback ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token от UOM. Claims: sub (user_id), org (organization_id), role.
...
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
...
    AuthTokens:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token (TTL 15 минут)
        refresh_token:
          type: string
          description: Refresh token (TTL 30 дней)
        expires_in:
          type: integer
          description: Время жизни access token в секундах

    # ---- Contracts ----
    DocumentStatus:
      type: string
      enum: [ACTIVE, ARCHIVED, DELETED]
...